requirements.txt
decorator==4.0.9
Django==1.9.4
django-tastypie==0.13.3
future==0.15.0
geopy==1.11.0
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pytest==2.8.7
python-dateutil==2.5.0
python-mimeparse==1.5.1
pytz==2015.7
PyYAML==3.11
requests==2.7.0
requests-mock==0.6.0
simplegeneric==0.8.1
six==1.9.0
sodapy==1.2.0
traitlets==4.2.1
